Added python version condition when delete packages

+++ building/lambda/build-lambda-layer.sh
@@ -85,10 +85,14 @@ pip3 install . -t ./python
 rm -rf python/pyarrow*
 rm -rf python/boto*
 rm -rf python/setuptools*
+
+if [ $(echo ${1} |grep -Po 'py3.\K[\d+]*') -gt 7 ] ;
+then
 rm -rf python/s3transfer*
 rm -rf python/six*
 rm -rf python/urllib3*
 rm -rf python/jmespath*
+fi
 
 rm -f /aws-sdk-pandas/dist/pyarrow_files/pyarrow/libarrow.so
 rm -f /aws-sdk-pandas/dist/pyarrow_files/pyarrow/libparquet.so
